refactor(legacy): move wait_random into separate file
